TB-105919, TB-128919: work around ABI break in fs::path
There is an ABI break in the std::filesystem::path between GCC 8 and
GCC 9 (in 4f87bb8d6e8, to be more precise) -- a private field was
changed along with some implementation details.

Unfortunately, since filesystem library implementation was still
deemed unstable in GCC 8, the implementors didn't think it was
necessary to handle this break gracefully. As a result, if you link an
application with libstdc++fs from GCC 8, but run it with libstdc++
from GCC 9+, there's a risk that filesystem functions with a different
ABI requirements would be called on your objects.

This patch closes the gap by adding all affected std::filesystem
classes in an inline namespace and assigning an ABI tag to it. This
way, we're making sure that filesystem symbols from GCC 8 and GCC 9+
are different on a linkage level, satisfying ODR requirements and
making sure that all function calls will always resolve to the
versions of the library they were compiled/linked with.

This patch, hopefully, will allow us to safely mix components compiled
with different libstdc++ versions in one process, assuming everything
else that we use from libstdc++ has stable ABI.

The patch is included in the libstdc++ distribution shipped with SDK
package starting with Tbricks 2.27.